Added possible configuration of treshold. cleaned up code to be more consistent

+++ main/include/utils.h
@@ -5,7 +5,7 @@ typedef struct {
     float temperature;
     float soil_moisture;
     float relative_humidity;
-    int light;
+    float light;
 } measurements_t;
 
 #endif
